minitalk

+++ minitalk/include/minitalk.h
@@ -6,7 +6,7 @@
 /*   By: cbouhadr <cbouhadr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/09 10:34:33 by cbouhadr          #+#    #+#             */
-/*   Updated: 2024/12/09 11:28:03 by cbouhadr         ###   ########.fr       */
+/*   Updated: 2024/12/09 14:41:45 by cbouhadr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
 #include <assert.h>
 #include <signal.h>
 
-size_t	ft_strlen(char *str);
+int	    ft_strlen(char *str);
 char    *ft_strjoin_char(char *str, char c);
 void    ft_print_bit(int n);
 void	ft_send_int_to_pid(int n, int pid, int ms);
